refactor: Bridge 클래스 제거
Bridge가 래핑하는 불리언 리스트가 Line과 동일하다고 판단하여 제거하고 기능을 Line으로 옮겼습니다.
